+++ core-codemods/src/main/java/io/codemodder/codemods/CreateConstantForLiteral.java
@@ -13,7 +13,7 @@ import com.github.javaparser.ast.nodeTypes.NodeWithSimpleName;
 import com.github.javaparser.ast.type.ClassOrInterfaceType;
 import com.github.javaparser.ast.type.Type;
 import io.codemodder.CodemodInvocationContext;
-import triage.Issue;
+import io.codemodder.sonar.model.Issue;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
