+++ core-codemods/src/main/java/io/codemodder/codemods/DefineConstantForLiteral.java
@@ -10,9 +10,9 @@ import io.codemodder.CodemodInvocationContext;
 import io.codemodder.Position;
 import io.codemodder.RegionNodeMatcher;
 import io.codemodder.SourceCodeRegion;
-import triage.Flow;
-import triage.Issue;
-import triage.TextRange;
+import io.codemodder.sonar.model.Flow;
+import io.codemodder.sonar.model.Issue;
+import io.codemodder.sonar.model.TextRange;
 import java.util.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
