+++ core-codemods/src/main/java/io/codemodder/codemods/DefineConstantForLiteralCodemod.java
@@ -8,7 +8,7 @@ import io.codemodder.javaparser.ChangesResult;
 import io.codemodder.providers.sonar.ProvidedSonarScan;
 import io.codemodder.providers.sonar.RuleIssues;
 import io.codemodder.providers.sonar.SonarPluginJavaParserChanger;
-import triage.Issue;
+import io.codemodder.sonar.model.Issue;
 import javax.inject.Inject;
 
 /** A codemod for defining a constant for a literal string that is duplicated n times. */
