+++ core-codemods/src/main/java/io/codemodder/codemods/OverridesMatchParentSynchronizationCodemod.java
@@ -10,7 +10,7 @@ import io.codemodder.javaparser.ChangesResult;
 import io.codemodder.providers.sonar.ProvidedSonarScan;
 import io.codemodder.providers.sonar.RuleIssues;
 import io.codemodder.providers.sonar.SonarPluginJavaParserChanger;
-import triage.Issue;
+import io.codemodder.sonar.model.Issue;
 import java.util.Optional;
 import javax.inject.Inject;
