+++ core-codemods/src/main/java/io/codemodder/codemods/RemoveUselessParenthesesCodemod.java
@@ -9,7 +9,7 @@ import io.codemodder.javaparser.ChangesResult;
 import io.codemodder.providers.sonar.ProvidedSonarScan;
 import io.codemodder.providers.sonar.RuleIssues;
 import io.codemodder.providers.sonar.SonarPluginJavaParserChanger;
-import triage.Issue;
+import io.codemodder.sonar.model.Issue;
 import javax.inject.Inject;
 
 /** Codemod to remove useless pair of parentheses */
