+++ core-codemods/src/test/resources/test-applications/move-switch-default-last/src/main/java/ai/pixee/integration/moveswitchdefaultlast/MoveSwitchDefaultLastApplication.java
@@ -1,4 +1,4 @@
-package ai.pixee.integration.moveswitchdefaultlast;
+package io.pixee.integration.moveswitchdefaultlast;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
