+++ core-codemods/src/test/resources/test-applications/move-switch-default-last/src/main/java/ai/pixee/integration/moveswitchdefaultlast/api/TestController.java
@@ -1,4 +1,4 @@
-package ai.pixee.integration.moveswitchdefaultlast.api;
+package io.pixee.integration.moveswitchdefaultlast.api;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
