+++ core-codemods/src/test/resources/test-applications/replace-apache-defaulthttpclient/src/main/java/ai/pixee/integration/replaceapachedefaulthttpclient/api/TestController.java
@@ -1,4 +1,4 @@
-package ai.pixee.integration.replaceapachedefaulthttpclient.api;
+package io.pixee.integration.replaceapachedefaulthttpclient.api;
 
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.HttpGet;
