+++ core-codemods/src/test/resources/test-applications/verbose-request-mapping/src/main/java/ai/pixee/integration/verboserequestmapping/VerboseRequestMappingApplication.java
@@ -1,4 +1,4 @@
-package ai.pixee.integration.verboserequestmapping;
+package io.pixee.integration.verboserequestmapping;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
