+++ sonar-api-model/src/main/java/io/codemodder/sonar/model/Issue.java
@@ -1,4 +1,4 @@
-package triage;
+package io.codemodder.sonar.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
